// types.ts
export type UpdateCampaignBudget = {
  resource_name?: string;
  explicitly_shared?: boolean;
  amount_micros?: number | string;
  delivery_method?: enums.BudgetDeliveryMethod;
};